==> jupyter_notebooks/01 - Data Collection.ipynb <==
# **Data Collection**

## Objectives

* Install dependencies
* Fetch data from Kaggle
* Do an initial inspection of the data
* Save the raw data in the outputs folder

### Inputs

* Kaggle JSON file - used for authentification

### Outputs

* Raw dataset, saved in the output folder as a csv file


---

## Set up the Working Directory

Define and confirm the current working directory

import os
current_dir = os.getcwd()
os.chdir(os.path.dirname(current_dir))
current_dir = os.getcwd()
current_dir

## Install Python Packages

Install dependencies from requirements.txt

%pip install -r requirements.txt

___

## Fetch Dataset from Kaggle

Install Kaggle

%pip install kaggle==1.5.12

Add the Kaggle authentication token to the root directory.
Read more in the [Kaggle Docs](https://github.com/Kaggle/kaggle-api).

Recognise the JSON token in this session.

os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

Define the Kaggle dataset and destination folder.
Download the the dataset.

KaggleDatasetPath = "dhanushnarayananr/credit-card-fraud"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

Unzip the downloaded file, delete the zip file and delete the kaggle.json file

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

___

## Initial Inspection of Data

import pandas as pd
df = pd.read_csv(f"inputs/datasets/raw/card_transdata.csv")
df.head()

df.info()

Some variables (repeat_retailer, used_chip used_pin_number, online_order & fraud) are encoded as binary, with 1 representing True and 0 representing False. However, they are currently stored as floats. To save memory and add clarity to their binary nature, these will be converted to integers.

Credit: [Memory Efficient Data Science: Types](https://towardsdatascience.com/memory-efficient-data-science-types-53423d48ba1d)

binary_columns = ['repeat_retailer','used_chip','used_pin_number','online_order','fraud']

for column in binary_columns:
    df[column] = df[column].astype('uint8')

df.head()

___

## Save Files to Repo

import os
try:
  os.makedirs(name=f'outputs/datasets/collection')
except Exception as e:
  print(e)

df.to_csv(f"outputs/datasets/collection/card_transactions.csv",index=False)
